Add CompositionPlayer
What if the user starts typing a player name before making any hero selections?

We want to save "zioncross" as the first player for that comp, but the only thing tying a player to a composition is the player_selections table, which requires a map segment and a hero selection.

No longer!